feat: Encode tokens when proxying downloads
When the proxy download endpoint is called,
Terraform will not include any access token.
In cases when Orbit is running without a default
access token, that means all requests will result
in a 404 from Github.

This changes the behaviour so that whenever a token
is explicitly passed to the Downloads endpoint (i.e.
the one giving Terraform the URL to actually
download - the proxy endpoint), it encodes the token
using AES and adds it as a query parameter to the
URL returned.

That way the proxy endpoint can decrypt the token
from the URL and use it when calling Github. The
validity fo the encrypted token is 60 seconds by
default.

Right now the AES secret is passed as a byte slice
in the environment variables (i.e. it must be
base64 encoded), and it is expected to be either
16, 24, or 32 bytes long.